requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
mail-parser==3.15.0
mysqlclient==2.1.0
pytz==2021.3
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
django-environ